remove 31-bit s390 build

+++ update-image.sh
@@ -7,7 +7,7 @@ docker build -q -t $IMAGE .
 
 NEW_IMAGE=$(docker images -q $IMAGE)
 
-if [[ $OLD_IMAGE != $NEW_IMAGE ]] ; then
+if [[ $OLD_IMAGE != "" ]] && [[ $OLD_IMAGE != $NEW_IMAGE ]] ; then
   echo "docker image $OLD_IMAGE has been updated as $NEW_IMAGE"
   docker rmi $OLD_IMAGE
 fi
